feat: move some dummy test functions into test_utils so we can use it in other places
