+++ rust/main/agents/relayer/src/test_utils/mod.rs
@@ -1,3 +1,4 @@
+pub mod dummy_data;
 pub mod mock_aggregation_ism;
 pub mod mock_base_builder;
 pub mod mock_ism;
